Bank Refactoring with Decorators (#28)
This PR refactors the bank application to use Operon
transaction/workflow/communicator/endpoint decorators and Operon HTTP
server. This results in much simpler code.

In this PR, I also temporarily removed JWT token authentication and
authorization. I will add it back in the future once Operon server
supports customized middleware.

One issue: I put the `BankEndpoints` class (with decorated endpoints) in
a separate file and import it from the main.ts file. However, the
decorators would not run at all unless we use this class. As a temporary
solution, I created an empty static `load()` function and call it from
the main.ts file. In the future, we will likely have a better way to
register a specific class/file (@devhawk is working on this).