Adjust socialts to new operon, add UI (#30)
This brings a copy of the UI into the operon app; needed for coevolution
of frontend and backend.
Updated to match operon's registration.
I would get rid of the external registration once auth is pluggable...
auth is the one thing this currently does that is not in operon.
This has the beginning of an S3 upload, but that needs to be turned into
a communicator in upcoming PRs. Don't want to lose the basic research
though.