Bank: add auth middleware (#34)
This PR adds `authenticate` middleware for the bank demo that maps the
claimed roles in JWT token to specified roles in Operon. Also adds a
basic Github action for bank so we can make sure it compiles and passes
the linter before merging.

Some observation for the authentication middleware (added TODOs in the
code as well):
- Can we support class-wide decorators? Now I have to manually add
`@RequiredRole` on top of every method.
- Should OperonHTTPAuthMiddleware be a function type not an interface?
The semantics here is a bit verbose. we must pass in `authMiddleware:
{authenticate: <actual function>}`. Wonder if we can just have a type
called `authMiddleware`?
- The authenticate middleware requires directly modifying the
HandlerContext and returns a boolean, which could be a bit tricky to use
(e.g., which variables it should and shouldn't set?). I think a better
semantics would be to return an `authenticatedReturn` object if success
and throw an error if the request is not authenticated.
- Should we support a permission-based model?

Nits:
- In Github actions, Prisma doesn't pick up the `.env` file and we have
to explicitly set `DATABASE_URL` as an env variable.