Bank use middleware decorators (#46)
This PR updates bank to use the latest middleware decorators in Operon
v0.0.4.

In general it works smoothly, but we need to be very careful about
circular dependencies as it can cause the middleware decorator input to
be `undefined`. That's the reason we have to declare bank middleware
functions in a separate file here.