Merge payment frontend into backend (#124)
The `ejs` code is now hosted in the backend; this simplifies deployment
of payment significantly and shop overall somewhat.

There is no longer any need for the OpenAPI client for payment, so I
took out the generator. The generator lives on in shop-backend for use
by shop-frontend, which is a sufficient illustration of how it works.

Slight changes to warning messages and comments in shop.

Stabilization of unit test - fix a timing hole.

Up SDK version to 0.11.9. (Do that to other apps as well, apply audit
fixes, to prevent the frustrations seen from affecting other apps.)

README section on cloud deployment and monitoring dashboard.