Widget Store update (#179)
This revamps the frontend of the widget store, adding two side-panels:
1. Shop internals. Shows a table for the current inventory, as well as a
table for all recorded orders, if existing. Orders are updated in
real-time.
2. Server Tools. The "Crash the Application" button, with some
explanatiory text.

Some backend changes were made to accommodate this, e.g., adding the
scheduled workflow that "does work" on each order to bring it from PAID
to DISPATCHED.

Unfortunately, due to some
[issues](https://dbosinc.slack.com/archives/C05CDE2GR2P/p1724186595181399)
of serving static files (and referencing them among themselves), all
frontend code is in one big `app.html`. This is not ideal, but allows us
to "just serve" the frontend via DBOS. In a real-world project, one
would probably separate the frontend into its own app that could be
deployed independently. In order to make this a single-deploy DBOS demo
app, the current approach was chosen.

To play around, the app is deployed here:
https://leanderneiss-widget-store.cloud.dbos.dev/