TypeScript v2 Express starter (#240)
This PR adds a TypeScript equivalent of the `dbos-app-starter`. It has
the exact same functionality as the Python one: launch a background
task, crash it, and then watch it resumes from where it left off.

Also fixed the console link in the Python starter.

Note: The log level specification is quite different from Python. I have
to use lower case `'info'` instead of `INFO`.