update node toolbox demo app (#327)
Similar to what #325 did for the starter app, but more complex as this
app has a data source and uses the app db.

In addition to #325 style changes:
* replaces `knexfile.js` with `setup-db.mts`. Knex file runs migrations
but does not create the database if it doesn't exist. `setup-db` does
both.
* added comments around the use of `DBOS_DATABASE_URL` to help reader
understand this is used to enable DBOS Cloud depoyment and that self
hosted apps can use whatever mechanism they wish for managing connection
info
* makes mention of feasibility running sys and app db on separate
servers, but doesn't explain how in order to make the readme simpler